Fixed validatePartTimeSalary tests
